code smell fix - Update this function so that its implementation is not identical to the one
