+++ package.json
@@ -19,6 +19,7 @@
   "lint-staged": {
     "**/*.{js,jsx}": [
       "npm run lint:fix --prefix server",
+      "npm run test:staged --prefix server",
       "git add ."
     ],
     "**/*.{ts,tsx}": [
